+++ src/js/App.js
@@ -1,9 +1,11 @@
-import React from 'libs/react/react.min';
-import ReactDom from 'libs/react/react-dom.min';
+import React from 'react';
+import ReactDOM from 'react-dom';
 
-export default class App extends React.Component {
-    constructor() {
-        super();
-        console.log('Hello from App');
+export default class App extends React.Component {    
+    render() {
+        ReactDOM.render(
+            <span>Hello from React!</span>
+            , document.getElementById('app')
+        );
     }
 }